Cambios 4

+++ apirestelevadores/src/main/java/com/maticolque/apirestelevadores/controller/DestinoController.java
@@ -1,14 +1,18 @@
 package com.maticolque.apirestelevadores.controller;
 
+import com.maticolque.apirestelevadores.dto.ErrorDTO;
 import com.maticolque.apirestelevadores.dto.RespuestaDTO;
 import com.maticolque.apirestelevadores.model.Destino;
 import com.maticolque.apirestelevadores.service.DestinoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+
 
 @RestController
 @RequestMapping("api/v1/destino")
@@ -20,38 +24,93 @@ public class DestinoController {
 
     //GET
     @GetMapping
-    public List<Destino> listarTodo(){
-        return destinoService.getAllDestino();
+    public ResponseEntity<?> listarTodo() {
+        try {
+            List<Destino> destinos = destinoService.getAllDestino();
+
+            if (destinos.isEmpty()) {
+                // Crear instancia de ErrorDTO con el código de error y el mensaje
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("La base de datos está vacía, no se encontraron destinos.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
+            }
+
+            return ResponseEntity.ok(destinos);
+        } catch (Exception e) {
+            // Crear instancia de ErrorDTO con el código de error y el mensaje
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("ERR_INTERNAL_SERVER_ERROR")
+                    .message("Error al obtener la lista de destinos: " + e.getMessage())
+                    .build();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorDTO);
+        }
     }
+
+
 
     //GET POR ID
     @GetMapping("/{id}")
-    public Destino buscarDestinoPorId(@PathVariable Integer id)
-    {
-        return destinoService.buscarDestinoPorId(id);
+    public ResponseEntity<?> buscarDestinoPorId(@PathVariable Integer id) {
+        try {
+            Destino destinoExistente = destinoService.buscarDestinoPorId(id);
+
+            if (destinoExistente == null) {
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("El ID que intenta buscar no existe.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
+
+            } else {
+                return ResponseEntity.ok(destinoExistente);
+            }
+        } catch (Exception e) {
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("ERR_INTERNAL_SERVER_ERROR")
+                    .message("Error al buscar el destino. " + e.getMessage())
+                    .build();
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, errorDTO.toString(), e);
+        }
     }
 
     //POST
     @PostMapping
-    public RespuestaDTO<Destino> crearMedioElevacion(@RequestBody Destino destino){
+    public RespuestaDTO<Destino> crearMedioElevacion(@RequestBody Destino destino) {
+        try {
+            // Realizar validación de los datos
+            if (destino.getDst_codigo().isEmpty() || destino.getDst_detalle().isEmpty()) {
+                throw new IllegalArgumentException("Todos los datos de destino son obligatorio.");
+            }
 
-
-        Destino nuevoDestino = destinoService.createDestino(destino);
-        return new RespuestaDTO<>(nuevoDestino, "Destino creado con éxito");
-
+            // Llamar al servicio para crear el destino
+            Destino nuevoDestino = destinoService.createDestino(destino);
+            return new RespuestaDTO<>(nuevoDestino, "Destino creado con éxito.");
+        } catch (IllegalArgumentException e) {
+            // Capturar excepción de validación
+            return new RespuestaDTO<>(null, "Error al crear un nuevo destino: " + e.getMessage());
+        } catch (Exception e) {
+            // Capturar otras excepciones
+            e.printStackTrace();
+            return new RespuestaDTO<>(null, "Error al crear un nuevo destino: " + e.getMessage());
+        }
     }
 
     //PUT
     @PutMapping("editar/{id}")
     //@ResponseStatus(HttpStatus.OK) // Puedes usar esta anotación si solo quieres cambiar el código de estado HTTP
-    public ResponseEntity<String> actualizarDestino(@PathVariable Integer id, @RequestBody Destino destino) {
+    public ResponseEntity<?> actualizarDestino(@PathVariable Integer id, @RequestBody Destino destino) {
         try {
             // Lógica para modificar el medio de elevación
             Destino destinoExistente = destinoService.buscarDestinoPorId(id);
 
             if (destinoExistente == null) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                        .body("No se encontró el Destino con el ID proporcionado");
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("El ID que intenta modificar no existe.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
             }
 
             //Modificar valores
@@ -59,22 +118,52 @@ public class DestinoController {
             destinoExistente.setDst_detalle(destino.getDst_detalle());
             destinoExistente.setDst_activo(destino.isDst_activo());
 
-            // Agrega más propiedades según tu modelo
             destinoService.updateDestino(destinoExistente);
 
-            return ResponseEntity.ok("La modificación se ha realizado correctamente");
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("200 OK")
+                    .message("La modificación se ha realizado correctamente.")
+                    .build();
+            return ResponseEntity.status(HttpStatus.OK).body(errorDTO);
 
         } catch (Exception e) {
             // Manejar otras excepciones no específicas y devolver un código y mensaje genéricos
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body("Se produjo un error al intentar modificar el Destino");
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("404 NOT FOUND")
+                    .message("Error al modificar el Destino."+ e.getMessage())
+                    .build();
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorDTO);
+
         }
     }
 
 
     //DELETE
     @DeleteMapping("eliminar/{id}")
-    public void elimimarDestino(@PathVariable Integer id){
-        destinoService.deleteDestinoById(id);
+    public ResponseEntity<?> elimimarDestino(@PathVariable Integer id) {
+        try {
+            Destino destinoExistente = destinoService.buscarDestinoPorId(id);
+
+            if (destinoExistente == null) {
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("El ID que intenta eliminar no existe.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
+            } else {
+                destinoService.deleteDestinoById(id);
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("200 OK")
+                        .message("Destino eliminado correctamente.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.OK).body(errorDTO);
+            }
+        } catch (DataAccessException e) { // Captura la excepción específica de acceso a datos
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("ERR_INTERNAL_SERVER_ERROR")
+                    .message("Error al eliminar el destino. " + e.getMessage())
+                    .build();
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, errorDTO.toString(), e);
+        }
     }
 }
