+++ apirestelevadores/src/main/java/com/maticolque/apirestelevadores/controller/DistritoController.java
@@ -1,14 +1,18 @@
 package com.maticolque.apirestelevadores.controller;
 
+import com.maticolque.apirestelevadores.dto.ErrorDTO;
 import com.maticolque.apirestelevadores.dto.RespuestaDTO;
 import com.maticolque.apirestelevadores.model.Distrito;
 import com.maticolque.apirestelevadores.service.DistritoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping("api/v1/distrito")
@@ -19,38 +23,92 @@ public class DistritoController {
 
     //GET
     @GetMapping
-    public List<Distrito> listarTodo(){
-        return distritoService.getAllDistrito();
+    public ResponseEntity<?> listarTodo() {
+        try {
+            List<Distrito> distrito = distritoService.getAllDistrito();
+
+            if (distrito.isEmpty()) {
+                // Crear instancia de ErrorDTO con el código de error y el mensaje
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("La base de datos está vacía, no se encontraron Distritos.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
+            }
+
+            return ResponseEntity.ok(distrito);
+
+        } catch (Exception e) {
+            // Crear instancia de ErrorDTO con el código de error y el mensaje
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("ERR_INTERNAL_SERVER_ERROR")
+                    .message("Error al obtener la lista de Distrito: " + e.getMessage())
+                    .build();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorDTO);
+        }
     }
+
 
     //GET POR ID
     @GetMapping("/{id}")
-    public Distrito buscarDestinoPorId(@PathVariable Integer id)
-    {
-        return distritoService.buscarDistritoPorId(id);
-    }
+    public ResponseEntity<?> buscarDistritoPorId(@PathVariable Integer id) {
+        try {
+            Distrito distritoExistente = distritoService.buscarDistritoPorId(id);
 
+            if (distritoExistente == null) {
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("El ID que intenta buscar no existe.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
+
+            } else {
+                return ResponseEntity.ok(distritoExistente);
+            }
+        } catch (Exception e) {
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("ERR_INTERNAL_SERVER_ERROR")
+                    .message("Error al buscar el distrito. " + e.getMessage())
+                    .build();
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, errorDTO.toString(), e);
+        }
+    }
     //POST
     @PostMapping
-    public RespuestaDTO<Distrito> crearMedioElevacion(@RequestBody Distrito distrito){
+    public RespuestaDTO<Distrito> crearDistrito(@RequestBody Distrito distrito) {
+        try {
+            // Realizar validación de los datos
+            if (distrito.getDis_codigo() == null || distrito.getDis_nombre().isEmpty()) {
+                throw new IllegalArgumentException("Todos los datos de destino son obligatorio");
+            }
 
+            // Llamar al servicio para crear el destino
+            Distrito nuevoDestino = distritoService.createDistrito(distrito);
+            return new RespuestaDTO<>(nuevoDestino, "Distrito creado con éxito");
 
-        Distrito nuevoDistrito = distritoService.createDistrito(distrito);
-        return new RespuestaDTO<>(nuevoDistrito, "Distrito creado con éxito");
+        } catch (IllegalArgumentException e) {
+            // Capturar excepción de validación
+            return new RespuestaDTO<>(null, "Error al crear un nuevo distrito: " + e.getMessage());
 
+        } catch (Exception e) {
+            return new RespuestaDTO<>(null, "Error al crear un nuevo distrito: "  + e.getMessage());
+        }
     }
 
     //PUT
     @PutMapping("editar/{id}")
     //@ResponseStatus(HttpStatus.OK) // Puedes usar esta anotación si solo quieres cambiar el código de estado HTTP
-    public ResponseEntity<String> actualizarDestino(@PathVariable Integer id, @RequestBody Distrito distrito) {
+    public ResponseEntity<?> actualizarDistrito(@PathVariable Integer id, @RequestBody Distrito distrito) {
         try {
             // Lógica para modificar el medio de elevación
             Distrito destinoExistente = distritoService.buscarDistritoPorId(id);
 
             if (destinoExistente == null) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                        .body("No se encontró el Distrito con el ID proporcionado");
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("El ID que intenta modificar no existe.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
             }
 
             //Modificar valores
@@ -61,19 +119,50 @@ public class DistritoController {
             // Agrega más propiedades según tu modelo
             distritoService.updateDistrito(destinoExistente);
 
-            return ResponseEntity.ok("La modificación se ha realizado correctamente");
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("200 OK")
+                    .message("La modificación se ha realizado correctamente.")
+                    .build();
+            return ResponseEntity.status(HttpStatus.OK).body(errorDTO);
+
 
         } catch (Exception e) {
             // Manejar otras excepciones no específicas y devolver un código y mensaje genéricos
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body("Se produjo un error al intentar modificar el Distrito");
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("404 NOT FOUND")
+                    .message("Error al modificar el Distrito."+ e.getMessage())
+                    .build();
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorDTO);
         }
     }
 
 
     //DELETE
     @DeleteMapping("eliminar/{id}")
-    public void elimimarDestino(@PathVariable Integer id){
-        distritoService.deleteDistritoById(id);
+    public ResponseEntity<?> elimimarDistrito(@PathVariable Integer id) {
+        try {
+            Distrito distritoExistente = distritoService.buscarDistritoPorId(id);
+
+            if (distritoExistente == null) {
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("El ID que intenta eliminar no existe.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
+            } else {
+                distritoService.deleteDistritoById(id);
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("200 OK")
+                        .message("Distrito eliminado correctamente.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.OK).body(errorDTO);
+            }
+        } catch (DataAccessException e) { // Captura la excepción específica de acceso a datos
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("ERR_INTERNAL_SERVER_ERROR")
+                    .message("Error al eliminar el Distrito. " + e.getMessage())
+                    .build();
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, errorDTO.toString(), e);
+        }
     }
 }
