+++ apirestelevadores/src/main/java/com/maticolque/apirestelevadores/controller/EmpresaController.java
@@ -1,14 +1,18 @@
 package com.maticolque.apirestelevadores.controller;
 
+import com.maticolque.apirestelevadores.dto.ErrorDTO;
 import com.maticolque.apirestelevadores.dto.RespuestaDTO;
 import com.maticolque.apirestelevadores.model.Empresa;
 import com.maticolque.apirestelevadores.service.EmpresaService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping("api/v1/empresa")
@@ -20,36 +24,94 @@ public class EmpresaController {
 
     //GET
     @GetMapping
-    public List<Empresa> listarTodo(){
-        return empresaService.getAllEmpresa();
+    public ResponseEntity<?> listarTodo() {
+        try {
+            List<Empresa> empresas = empresaService.getAllEmpresa();
+
+            if (empresas.isEmpty()) {
+                // Crear instancia de ErrorDTO con el código de error y el mensaje
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("La base de datos está vacía, no se encontraron Empresas.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
+            }
+
+            return ResponseEntity.ok(empresas);
+
+        } catch (Exception e) {
+            // Crear instancia de ErrorDTO con el código de error y el mensaje
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("ERR_INTERNAL_SERVER_ERROR")
+                    .message("Error al obtener la lista de Empresas: " + e.getMessage())
+                    .build();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorDTO);
+        }
     }
 
 
     //GET POR ID
     @GetMapping("/{id}")
-    public Empresa buscarEmpresaPorId(@PathVariable Integer id)
-    {
-        return empresaService.buscarEmpresaPorId(id);
+    public ResponseEntity<?> buscarEmpresaPorId(@PathVariable Integer id) {
+        try {
+            Empresa empresaExistente = empresaService.buscarEmpresaPorId(id);
+
+            if (empresaExistente == null) {
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("El ID que intenta buscar no existe.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
+
+            } else {
+                return ResponseEntity.ok(empresaExistente);
+            }
+        } catch (Exception e) {
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("ERR_INTERNAL_SERVER_ERROR")
+                    .message("Error al buscar la Empresa. " + e.getMessage())
+                    .build();
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, errorDTO.toString(), e);
+        }
     }
 
     //POST
     @PostMapping
-    public RespuestaDTO<Empresa> crearEmpresa(@RequestBody Empresa empresa){
-        Empresa nuevaEmpresa= empresaService.createEmpresa(empresa);
-        return new RespuestaDTO<>(nuevaEmpresa, "Empresa creada con éxito");
+    public RespuestaDTO<Empresa> crearEmpresa(@RequestBody Empresa empresa) {
+        try {
+            // Realizar validación de los datos
+            if (empresa.getEmp_cuit().isEmpty() || empresa.getEmp_telefono().isEmpty() || empresa.getEmp_razon().isEmpty() ||
+                    empresa.getEmp_domic_legal().isEmpty() || empresa.getEmp_correo().isEmpty()) {
+                throw new IllegalArgumentException("Todos los datos de la empresa son obligatorios");
+            }
+
+            // Llamar al servicio para crear el destino
+            Empresa nuevoEmpresa = empresaService.createEmpresa(empresa);
+            return new RespuestaDTO<>(nuevoEmpresa, "Empresa creado con éxito");
+
+        } catch (IllegalArgumentException e) {
+            // Capturar excepción de validación
+            return new RespuestaDTO<>(null, "Error al crear una nueva Empresa: " + e.getMessage());
+
+        } catch (Exception e) {
+            return new RespuestaDTO<>(null, "Error al crear una nueva Empresa: " + e.getMessage());
+        }
     }
 
     //PUT
     @PutMapping("editar/{id}")
     //@ResponseStatus(HttpStatus.OK) // Puedes usar esta anotación si solo quieres cambiar el código de estado HTTP
-    public ResponseEntity<String> actualizarEmpresa(@PathVariable Integer id, @RequestBody Empresa empresa) {
+    public ResponseEntity<?> actualizarEmpresa(@PathVariable Integer id, @RequestBody Empresa empresa) {
         try {
             // Lógica para modificar el medio de elevación
             Empresa empresaExistente = empresaService.buscarEmpresaPorId(id);
 
             if (empresaExistente == null) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                        .body("No se encontró la Empresa con el ID proporcionado");
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("El ID que intenta modificar no existe.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
             }
 
             //Modificar valores
@@ -60,21 +122,51 @@ public class EmpresaController {
             empresaExistente.setEmp_correo(empresa.getEmp_correo());
             empresaExistente.setEmp_activa(empresa.isEmp_activa());
 
-            // Agrega más propiedades según tu modelo
             empresaService.updateEmpresa(empresaExistente);
 
-            return ResponseEntity.ok("La modificación se ha realizado correctamente");
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("200 OK")
+                    .message("La modificación se ha realizado correctamente.")
+                    .build();
+            return ResponseEntity.status(HttpStatus.OK).body(errorDTO);
 
         } catch (Exception e) {
             // Manejar otras excepciones no específicas y devolver un código y mensaje genéricos
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body("Se produjo un error al intentar modificar la Empresa");
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("404 NOT FOUND")
+                    .message("Error al modificar la Empresa."+ e.getMessage())
+                    .build();
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorDTO);
         }
     }
 
+
     //DELETE
     @DeleteMapping("eliminar/{id}")
-    public void elimimarEmpresa(@PathVariable Integer id){
-        empresaService.deleteEmpresaById(id);
+    public ResponseEntity<?> elimimarEmpresa(@PathVariable Integer id) {
+        try {
+            Empresa empresaExistente = empresaService.buscarEmpresaPorId(id);
+
+            if (empresaExistente == null) {
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("404 NOT FOUND")
+                        .message("El ID que intenta eliminar no existe.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorDTO);
+            } else {
+                empresaService.deleteEmpresaById(id);
+                ErrorDTO errorDTO = ErrorDTO.builder()
+                        .code("200 OK")
+                        .message("Empresa eliminada correctamente.")
+                        .build();
+                return ResponseEntity.status(HttpStatus.OK).body(errorDTO);
+            }
+        } catch (DataAccessException e) { // Captura la excepción específica de acceso a datos
+            ErrorDTO errorDTO = ErrorDTO.builder()
+                    .code("ERR_INTERNAL_SERVER_ERROR")
+                    .message("Error al eliminar la Empresa. " + e.getMessage())
+                    .build();
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, errorDTO.toString(), e);
+        }
     }
 }
